Move TriplestorePopulator to cli package
